rename files and add missing namespace to payload conversion
