+++ gps_ubxm8_i2c/gps_ubxm8_i2c.cc
@@ -1,4 +1,4 @@
-#include "gps.h"
+#include "gps_ubxm8_i2c.h"
 
 #include <algorithm>
 #include <cmath>
@@ -128,7 +128,7 @@ void GpsUbxM8I2c::Reset() {
     _packetReader.reset();
 }
 
-UCIRP_GPS_PAYLOAD ConvertPayloadToECEF(UBX_NAV_PVT_PAYLOAD pvtPayload) {
+UCIRP_GPS_PAYLOAD GpsUbxM8I2c::ConvertPayloadToECEF(UBX_NAV_PVT_PAYLOAD pvtPayload) {
     UCIRP_GPS_PAYLOAD payload;
     payload.gpsFix = pvtPayload.fixType;
     payload.positionAcc = std::max((double)pvtPayload.hAcc / 1000, (double)pvtPayload.vAcc / 1000);
